requirements.txt
Flask>=2.3.2
pymongo>=4.7.2
python-dotenv>=1.0.1